authorization_service/requirements.txt
Flask==2.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==2.0.1
google-auth-oauthlib==0.4.6
google-cloud-secret-manager==2.19.0
python-dotenv==1.0.1
requests==2.26.0
gunicorn==22.0.0
google-api-python-client==2.31.0
